Organized remaining components
