+++ src/Routes.jsx
@@ -2,11 +2,11 @@ import Characters from './components/Characters/Characters';
 import Home from './components/Home';
 import NotFound from './components/NotFound';
 import Preferences from './components/Preferences';
-import SourceUpload from './components/SourceUpload';
+import SourceUpload from './components/Form/SourceUpload';
 import Encounter from './components/Encounters/Encounter';
 import Encounters from './components/Encounters/Encounters';
 import EncounterForm from './components/Encounters/EncounterForm';
-import Form from './components/Form';
+import Form from './components/Form/Form';
 import Monsters from './components/Monsters/Monsters';
 import Run from './components/Runs/Run';
 import Error from './components/Error';
