+++ src/components/Encounters/Encounters.jsx
@@ -1,4 +1,4 @@
-import List from '../List';
+import List from '../List/List';
 
 const Encounters = () => {
 	const config = {
