+++ src/components/Spells/Spell.jsx
@@ -1,4 +1,4 @@
-import CardView from '../CardView';
+import CardView from '../Card/CardView';
 import SpellCard from './SpellCard';
 
 const Spell = () => {
