+++ src/components/Spells/Spells.jsx
@@ -1,4 +1,4 @@
-import List from '../List';
+import List from '../List/List';
 
 const Spells = () => {
 	return <List title={'Spells'} storageKey={'spells'} canFilter paginate />;
